[Ide] Handle projects not always displayed when selecting template.
The New Project dialog supports project templates that are hidden
when File - New - Solution is selected but visible when Add - Add New
Project is selected. This causes problems when selecting the last
created project if one of these project templates is selected.

Previously if a project was created from a template that is only
displayed when adding a new project to a solution then the correct
project template would be selected on re-opeing the New Project
dialog only if adding a new project to the solution. When creating
a new solution the New Project dialog would not be able to find the
project template, since it is not shown, and would fall back to
selecting the first project template it can find.

Now with this commit the New Project dialog will save two last
created projects if the project template is only displayed when
adding a new project to an existing solution. So when opening the
New Project dialog to create a new solution and the previously created
project template is not visible it will show the last project that
was created when creating a new solution instead of selecting the
first project template it can find. If the project is visible in
both cases then only one project template is remembered by the New
Project dialog.