[Ide] Fix selected language not being updated in New Project dialog.
The selected language stored in the properties xml file would not
be updated unless the language drop down was selected. This would
mean that when a project template that did not support the last
language used and a new project was created the selected language
was not updated in the properties.