[NuGet] Fix build task not updated for Xamarin.Forms.
Fixed bug #32162 - Error initializing task
FixedCreateCSharpManifestResourceName when updating Forms NuGet package.
https://bugzilla.xamarin.com/show_bug.cgi?id=32162

If a solution that has multiple projects that reference Xamarin.Forms
is built once and then the NuGet package is updated the old MSBuild
task dll is still used when compiling.

To reproduce this bug:

1) Create a Xamarin.Forms project (PCL + Android + iOS).
2) Install Xamarin.Forms 1.4.3.6376 into each project.
3) Compile the solution.
4) Update Xamarin.Forms to version 1.4.4-pre3 or higher.
5) Rebuild the solution.
6) Build fails with:

Error initializing task FixedCreateCSharpManifestResourceName: Not
registered task FixedCreateCSharpManifestResourceName.

The FixedCreateCSharpManifestResourceName is a new task made available
in the newer version of Xamarin.Forms. The MSBuild host was still
using the old Xamarin.Forms.Build.Tasks.dll version 1.4.3. The NuGet
addin was only disposing the project builder for the project being
updated. If there are multiple projects in the solution the MSBuild
host would not be stopped so the old build task would still be used.
Now the NuGet addin will dispose the project builders for all projects
in the solution if an import was removed whilst updating the NuGet
package. This results in the MSBuild host being stopped and then
when the next build is done the MSBuild host is restarted and uses the
correct build task dll.

This is a change to the fix implemented in commit ac84628953fb45b87804c7c8112f52ff431e8a45