[NuGet] Do not show check for updates message in status bar.
Fixed bug #32670 - Do not show Checking for Updates message in status
bar.
https://bugzilla.xamarin.com/show_bug.cgi?id=32670

When a solution is opened the NuGet addin will check for package
updates. When it does this check a message is shown in the status bar
"Checking for package updates". From a user's point of view this seems
to indicate that they need to wait for the check to finish even though
this is not the case. So now the packages will be checked for updates
but a message will not be displayed in the status bar nor will any
messages be written to Package Console window.