[MacPlatform] Stop statusbar's popovers not going away
When the NSPopover appeared MainToolbar was receiving a NSWindowDidResizeNotification for the NSPopover, but it was causing the NSTrackingAreas in the StatusBar to be regenerated. This could cause a race condition when the old tracking area was
removed and the mouse exited the area before the new tracking area was installed.

Fixed the race condition by Only processing the NSWindowDidResizeNotification whenever it was the window that the statusbar was contained in.