[Git] Fix race condition in disposal of the repository.
Whenever there were still queries running, we'd be disposing
something in the subclass that would mess with the queries and
trigger an exception.

Let us check whether we're running any queries. If we're running,
just wait for them to end before disposing. Possibly a better fix
would be to use async and cancellation tokens, but the API is not
appropriate for that right now.

Bug 32471 - Shutdown exception from git addin