[AspNet] Fix web forms syntax mode crash.
Fixed bug #32979 - Can't enter quote (") character in script tag
https://bugzilla.xamarin.com/show_bug.cgi?id=32979

With an .aspx page that contains a script tag as shown below:

  <script type=></script>

Typing in a double quote character after the equals sign would cause
a crash which would terminate MonoDevelop.

The problem was that the information from the type attribute value is
used to find the mime type for the text inside the script tag. If the
mime type text cannot be found then the SyntaxMode would throw an
exception:

System.Exception: Rule mode:></script>
</body>
</html>

 not found in [Span: Color=, Rule=mode:></script>
</body>
</html>

 at Mono.TextEditor.Highlighting.SyntaxMode+ChunkParser.FoundSpanBegin (Mono.TextEditor.Highlighting.Span span, Int32 offset, Int32 length) in
  main/src/core/Mono.Texteditor/Mono.TextEditor.Highlighting/SyntaxMode.cs:561
 at MonoDevelop.AspNet.WebForms.WebFormsSyntaxMode+ASPNetSpanParser.ScanSpan (System.Int32& i) in
  main/src/addins/AspNet/WebForms/WebFormsSyntaxMode.cs:177
 at Mono.TextEditor.Highlighting.SyntaxMode+SpanParser.ParseSpans (Int32 offset, Int32 length) in
  main/src/core/Mono.Texteditor/Mono.TextEditor.Highlighting/SyntaxMode.cs:440
 at Mono.TextEditor.Highlighting.SyntaxMode+ChunkParser.GetChunks (Int32 offset, Int32 length) in
  main/src/core/Mono.Texteditor/Mono.TextEditor.Highlighting/SyntaxMode.cs:660
 at Mono.TextEditor.Highlighting.SyntaxMode+<GetChunks>c__Iterator0.MoveNext () in
  main/src/core/Mono.Texteditor/Mono.TextEditor.Highlighting/SyntaxMode.cs:132

This could also happen if the type attribute had an invalid mime type
whilst the script tag was inside a comment.

  <!--<script type="foo"></script>-->