[MacPlatform] Don't pass entry assembly exe files to OpenFiles
When Mono applications are run from the commandline or from within XS itself, the Cocoa system parses argv, stripping the executable name and any option arguments and passes the rest of the arguments to OpenFiles.

Unfortunately for mono applications (for example "mono-sgen MonoDevelop.exe") the executable name is interpreted as a filename and so OpenFiles receives it in the array.

We solve this by ignoring any filename that matches the Entry Assembly's location