[NuGet] Do not report standard NuGet errors to Insights.
Fixed bug #28145 - NuGet addin should not report NuGet errors to
Insights
https://bugzilla.xamarin.com/show_bug.cgi?id=28145

Standard errors, such as being unable to install a NuGet package into
a project that does not have a compatible target framework, were being
reported to Insights. These errors are reported in the Package Console
and logged locally but now they are not reported to Insights. Other
not important errors are also now no longer reported to Insights.