[NuGet] Support updating pre-release NuGet packages
Fixed bug #31071 - Xamarin Studio does not recognize pre-release
packages as updates
https://bugzilla.xamarin.com/show_bug.cgi?id=31071

Previously check for package updates would only show stable NuGet
package updates.

Check for package updates will now check for updates to pre-release
NuGet packages as well as stable packages. This involves potentially
two calls to nuget.org for each project if the project has stable
and pre-release NuGet packages. If a project only has stable NuGet
packages then only one call will be made to nuget.org to find
updates to the stable NuGet packages. Only if a NuGet package is
pre-release will pre-release updates be shown as available in the
Solution window.

Previously it was not possible to update a pre-release NuGet package
to a later pre-release  from the Solution window. If there was a later
stable NuGet package available then it was possible to update the
NuGet package. Previously the only way to update to a later
pre-release NuGet package was to use the Add Packages dialog.

Updating to a later pre-release NuGet package is now supported in the
Solution window. An individual pre-release NuGet package can be
updated by right clicking and selecting Update. When all packages in
a project or solution are updated if then pre-release NuGet packages
will be updated to later pre-release NuGet packages if they are
available.