[Core] Fix UI lock when building projects
The problem is that DotNetProjects made a call to GetReferencedAssemblies
each time a file is changed, and that method can be very slow the first time
is called (since it needs to run an msbuild target).

Instead of GetReferencedAssemblies(), we now call GetReferencedFileNames()
for each reference, which doesn't go through msbuild. It is not perfect
because in some cases may reference an assembly which is not explicitly
referenced in the project references, but it is good enough for now.

This should not be a problem in the new project model since there
GetReferencedAssemblies is asynchronous.