[Ide] Disable search widget background drawing
Setting the "transparent-bg-hint" flag inside widgets data
requests the theme engine not to fill the background of
the widget with its parents background color, making
widgets without own Gdk Window really transparent.

This works only if the engine supports this flag, which
is the case for many modern Gtk engines including
xamarin-gtk-theme.