These changes allow the language serve to report error when a partial failure occurs
Not all of the language service feature require that the parse and type
check results are both present.  So here we split the operation where
we can and allow errors to propagate on a type check failure which
results in an abort message fro the compiler.  Previously we would just
get no errors in the IDE and a spurious build failed during compile.