[MacPlatform] Rework Mac toolbar to use a single NSView
The Mac toolbar API makes it very hard to make a complex toolbar like we want, creating issues with centring widgets, adding and removing widgets, not updating sizes in real time and enforcing use of the overflow menu. All this leads to very complex and fragile code. Instead of using the NSToolbar API for everything, we create our own toolbar using NSViews and insert this NSView toolbar as the single toolbar item.