[Ide] Fix tasks not getting cleared from the statusbar
Due to project class abstraction changing, the task service would stop
going up the hierarchy on getting to a SolutionFolder. That was neither
a SolutionItem nor a WorkspaceItem, and it didn't match the Solution
itself, so it didn't pass the equality check.

Bug 27825 - [Roslyn] Warnings and Errors are not cleared from the status
bar when the solution / workspace is closed