[Core] Fix invalid warnings when archiving source files.
Fixed bug #30731 - Xamarin Studio shows error when creating packages.
https://bugzilla.xamarin.com/show_bug.cgi?id=30731

With a solution open, selecting Project - Create Archive, then
choosing to create an Archive of Sources package would show errors
in the Tool Output window about files being outside the solution or
project directory when they were not outside on the Mac.

The problem was that the solution/project files are copied to a
temp directory and the file name symlinks are fully resolved so the
path starts with "/private/var/" whilst the original temp directory
that was created starts with "/var/" so when comparing the paths the
files seemed to be in a different directory than the solution/project
directory. Now with this commit when this check is made the symlink
is resolved for the temp directory path so the comparison of the
paths now correctly determines if the files in the solution/project
are outside its directory.