[Ide] Support for CoreCompileDependsOn and Generators in project file items.
Context: https://mhut.ch/journal/2015/06/30/build_time_code_generation_msbuild

When a project is loaded, we look for targets defined in CoreCompileDependsOn and
the CompileItem results of these targets are included as source files into the
type system.

Whenever any file is changed that has a Generator value, we also run those targets
and notify the type system of changes in those extra included files. This is
performed by the existing MSBuildCustomTool processor.