[Core] Fix issue when saving multi-item includes
Fixed check on project items that was used to determine if an
item which was loaded from a wildcard or an include with multiple items
has been modified, in which case the multi-item include has to be
expanded. Also fixed logic to avoid expanding the include multiple times.

Fixes bug #35476 - Items are split and duplicated (incorrectly) when saving
project