[Windows] Open assemblies in assembly browser by default.
Fixed bug #35763 - Assemblies should open in assembly browser by
default
https://bugzilla.xamarin.com/show_bug.cgi?id=35763

Assemblies can be opened as a solution so this means the open file
dialog was selecting Solution Workbench as the default value in the
Open With drop down. To allow the assembly browser to be selected
as the default in the Open With drop down a check is made to see if
the files could be opened with the assembly browser and if so the
default selection is not set to be the Solution Workbench. This
allows the assembly browser to be selected by default.

Note that it is not currently possible to select the index of the
drop down. The IFileDialogCustomize.SetSelectedControlItem throws a
value out of range exception until ApplyControlPropertyChange is
called for the Items in the drop down. Setting the selected index
after the ApplyControlPropertyChange is called has the side effect of
visibly selecting the first item in the list and then changing
the selection to the correct item. So instead the order of the
items in the list is changed so the default one is first.