[Ide] Prompt to save files before opening New Project dialog.
Fixed bug #34974 - "New Solution" interacts poorly with
"save changes" dialog
https://bugzilla.xamarin.com/show_bug.cgi?id=34974

With unsaved files open and File - New Solution is selected after
trying to create a new solution then a save dialog is displayed for
each unsaved file. If this dialog is cancelled then the new solution
is not opened which is confusing. Also if the new project uses any
NuGet packages these will not be installed. The new solution has
been created but is not opened.

Now if there are unsaved files then the save dialog will appear before
the New Project dialog is displayed. If the save dialog is cancelled
then the New Project dialog is not opened. This prevents unsaved files
being open when a new solution is created.

Currently a save file dialog will appear for each unsaved file. This
is the current behavior when a solution is closed. It might be better
to instead show a dialog similar to the dialog when is shown when
the Quit menu is selected. This has a list of files that need saving
instead of a dialog for each file. If only one file needs saving then
the original save file dialog could be displayed.