[MacPlatform] Move the Modal window detection from a whitelist to a blacklist
Instead of playing whack-a-mole everytime a new version of OSX adds a new magic window that breaks our modal window detection, this change makes a blacklist of windows we know are modal.

Currently we only need to blacklist NSOpenPanel and NSSavePanel. If other window types are added to the code, they should be added to this list.

Fixes BXC #35465
Fixes BXC #34250