[Core] Fix cut and paste in solution pad not updating project file.
Fixed bug #35869 - Cut and paste file in Solution pad does not update
project file.
https://bugzilla.xamarin.com/show_bug.cgi?id=35869

With a project containing two folders foo and bar and an EmptyClass.cs
file in the bar folder then cutting and pasting the EmptyClass.cs from
bar to foo in the Solution pad the project file (.csproj) would not
be updated. The project file would still have 'bar\EmptyClass.cs' and
not 'foo\EmptyClass.cs'.

The problem was that the original ProjectFile was cloned and the
original UnevaluatedInclude property was copied. This contained the
old path to the file. If this property is set then it is used instead
of the Include property when saving the project file. The Include
property was correct but was not being used. So now when cloning the
ProjectFile the UnevaluatedInclude is cleared so the correct path
is written to the project file.