[Core] Fix copy and paste removing original file from project.
Copy a file from one folder and pasting it into another folder in
the Solution pad would result in the original file not being saved
in the project file (.csproj).

The problem was that the BackingItem was cloned when copying the
original ProjectFile so changes to the Include value were made to the
same underlying MSBuild BackItem. Now the BackingItem is cleared for
the cloned copy of the original file.