Fixed Bug 14907 - Extremely slow processing with very long line
This is a known problem with pango - that it becomes very slow with
large lines. However it's possible to cache some values that speed up
calculations and turning off syntax highlighting in that line helps a
lot as well.
My test case has become from barely usable to sluggish usable - which
is quite an improvement.