[Core] Allow empty item groups in MSBuild projects.
Previously when removing an item from a parent group using the
MSBuildProject the parent group would always be removed when it is
empty. The RemoveItem method now has a flag to allow this behaviour
to be overridden. This allows the empty item group to be kept in the
MSBuildProject. This can be useful when migrating projects so you can
remove all the existing items in the group and then add back items
to the same group without having to re-add a new group.