[AspNet] Fix MVC Views Configuration file template not being displayed
Fixed bug #36272 - ASP.NET MVC Views Configuration template is missing
https://bugzilla.xamarin.com/show_bug.cgi?id=36272

The ASP.NET MVC Views Configuration file template was using the
PermittedCreationPaths condition so it is only shown in the New File
dialog if the Views folder is selected in the Solution window and
there is no existing web.config file in this directory.

  <ParentProject PermittedCreationPaths="/Views" ExcludedFiles="Web.config" />

The PermittedCreationPaths started with a '/' so the paths being
compared for a project in directory /Projects/MyProject was:

  /Views == /Projects/MyProject/Views

Which never matched so the file template was not displayed. Removing
the leading '/' fixes the directory comparison so the file template
is now displayed.