[Ide] Fix crash in global search popup window.
Fixed bug #29879 - Crash if I type in search box
https://bugzilla.xamarin.com/show_bug.cgi?id=29879

If a search category throws an exception when its ItemCount is called
then this can cause the IDE to terminate. One way to reproduce this
with Xamarin Studio 5.10 or later is to install an old version of the
NuGet extensions addin, version 0.10 or earlier, then attempt to use
the global search in the top right of the IDE. This was causing an
exception to be thrown in the SearchPopupWindow's OnDrawContent
method when it tried to use the ItemCount property. This then
terminated the IDE.

Now a search category that throws an exception is removed from the
list of results so the OnDrawContent method never uses a badly
behaving search category.