[Core] Fix user preferences not loaded for workspace.
Fixed bug #30673 - [Roslyn] Configuration resets to first when loading monodevelop.sln and warning about translation project.
https://bugzilla.xamarin.com/show_bug.cgi?id=30673

The .userprefs file for a workspace (.mdw) was not being read when
the workspace was opened. This would cause the active configuration
in the main toolbar to always use the first available configuration,
usually Debug, since the user preferences were not loaded.