[CSharpBinding] Do not show search results from closed solution.
When a solution was opened and then closed the unified search would
still return results from the last opened solution since the
ProjectSearchCategory keeps the symbol information after the solution
is closed. This keeps alive the solution and Roslyn workspace but only
temporarily until the next solution is opened. However it does mean
that you can still search for types in the previously closed solution
before another solution is opened.

Now when the last workspace item is closed the symbol information is
cleared from the project search category. This prevents types being
returned in the search results when no solution is opened. It also
allows the solution and workspace to be garbage collected when the
solution is closed without having to wait for another solution to be
opened.