[Ide] Fix missing filename from save changes message.
Fixed bug #36561 - Typo in warning message.
https://bugzilla.xamarin.com/show_bug.cgi?id=36561

With an unsaved storyboard file, when File - New - Solution was
selected a message appears about saving changes before creating a
new solution. This message did not show the filename for storyboards.
Storyboards do not have the filename set for the document but instead
have the filename as the ViewContent.ContentName so this is now used
when displaying the message.