[Ide] Improve error reporting in New Project dialog.
The New Project dialog was always showing a message that the project
name had illegal characters even if the project was not due to the
project name. An invalid solution name or invalid location would also
produce the same error message is confusing. Now a more specific error
message will be displayed so it is possible to know whether the
solution or location is invalid.