[Ide] Fix toolbar behaviour when editing used out of solution
This fixes the toolbar popup not disappearing on clear icon click on
Windows.

This also fixes the fact that the popup didn't even go away when not
inside a solution, so it'd stay there for a long time.

Bug 36550 - [Windows-Toolbar] Search list window does not close when user
click on close(X) button appearing in search box