[IDE] Separate the notification model from the status bar
Previously to place a message in the statusbar, it had to be accessed directly. This commit introduces the NotificationService which is an attempt to separate out the data from the statusbar.
The NotificationService contains NotificationContexts. These contexts are the data in the model which the Statusbar displays. Sitting in between is the StatusbarContextHandler which marshals the data from the NotificationService and the statusbar listens for changes.

The NotificationService handles multiple contexts correctly, and the StatusbarContextHandler will switch between the contexts every 5 seconds.

This is an initial commit to discuss, it has only been tested on the OSX version of MD.