[Ide] Fix new PCL project added to solution not having any references.
Fixed bug #17234 - Adding a new PCL project to an existing solution
shows "empty" References folder before the solution is reloaded
https://bugzilla.xamarin.com/show_bug.cgi?id=17234

The problem was that when adding a new PCL project to an existing
solution the ITreeBuilder's current DataItem is the Project. When
loading a PCL project into the Solution window the ITreeBuilder's
current DataItem is the ProjectReferenceCollection. So the
ProjectReferenceFolderNodeBuilder was not finding the project since
it was not including the current DataItem in its search.