[Ide] Fix policies not being saved.
Fixed bug #11196 - Cannot change the code formatting policy
https://bugzilla.xamarin.com/show_bug.cgi?id=11196

When using the Apply Policy dialog, from the Project - Apply Policy
menu, changes made to the current project or solution policy were
not saved to the solution or project file so would only be active
during the current IDE session. Now the policy changes are saved
in the project or solution when the Apply Dialog is closed.