Fix msbuild log truncation issue
The msbuild output was sometimes truncated since it was sent from the builder
process to MD using a one-way WriteLine method, which means the call may return
before the remote method is executed. In some cases the remote logger was
disposed before the WriteLine method was completed, and that log was lost.

This patch removes the one-way method and adds code to periodically flush
the log to MD. Before returning the build result it ensures that all log has
been sent.

Fixes bug #35322 - Xamarin Studio diagnostic build output sometimes omits
some of the Task and Target "performance summary" lines