[Projects] Set the SolutionDir property when loading a project
The SolutionDir property is set by VS when loading a project. XS now does
the same. The value for SolutionDir is obtained from the solution context
provided when loading a project, so it will be available only one the project
is loaded as part of a solution. There were some cases in which the context
was not properly initialized (e.g. when reloading a project). This has also
been fixes.

This (and other previous commits) fixes bug #36700 - Project Import not
working for Xamarin Studio