[Ide] Ensure last selected category is on screen in New Project dialog
When the watchOS template category is displayed then the list of
categories does not fit on the screen. If the last selected project
template category was inside the Other top level category then the
category would be selected but it would be off screen. Now the
New Project dialog ensures the last selected category is on screen
by scrolling if necessary.