[Core] Fixed references not being saved to project file.
Fixed bug #36913 - Updating package with NuGet removed references
https://bugzilla.xamarin.com/show_bug.cgi?id=36913

If all references were removed from a project and then another added
then the reference would not be saved to the project file. One way to
reproduce this is to create a Xamarin.Forms project and then update
the Xamarin.Forms NuGet package in the PCL project. Another way is
to create a PCL project, add a reference to a file on disk, such as
Newtonsoft.Json.dll, remove the reference, then add the reference
again. However it is not a problem specific to a PCL project.

The problem was that when all the references are removed the parent
item group for the references is also removed. However this parent
item group remains in a bestGroups dictionary, which is used to group
similar items together, such as references. When a new reference was
then added it was added to the item group found in the bestGroups
dictionary, but that item group is not part of the MSBuildProject so
the references were not saved to disk.

Now when an item group parent is removed it is also removed from the
bestItemGroup dictionary.