[Ide] Fix project build status not updated when active config changed.
Fixed bug #34040 - Removing a project from a build configuration isn't
updated until refreshing the configuration
https://bugzilla.xamarin.com/show_bug.cgi?id=34040

With a solution containing two projects, if the solution options
dialog is opened, and in Build - Configurations - Configuration
Mappings one of the projects is then removed from building in the
currently active configuration by unchecking the its Build check box,
then on pressing OK on the dialog the Solution window will still
show the unchecked project as being built. If the active configuration
in the main toolbar is changed to another configuration and back
again the Solution window is updated and the project is then shown
as not being built in the current configuration by the project name
being shown in grey text.

The problem is that the ProjectNodeBuilder is not notified that the
build status has changed when a project is added or removed from
the current build configuration. Now when the modified solution
configurations are copied across to the solution on closing the
solution options dialog the Solution's ConfigurationsChanged event
fires. This event is used to trigger the ProjectNodeBuilder to
refresh the project build status.

The problem with this fix is that the event fires too often. It fires
several times on opening the solution options, and it fires several
times on closing the solution options dialog.