[NuGet] Fix status bar message changing too often on modifying packages.
Fixed bug #36918 - NuGet addin saves project after each change it makes
https://bugzilla.xamarin.com/show_bug.cgi?id=36918

The NuGet addin saves the project each time NuGet modifies the project.
This results in the "Saving..." message to briefly appear in the status
bar and causes the status bar message to appear to blink on and off.

To prevent the status bar message changing a different progress monitor
is now used when saving the project. This progress monitor does not
show the Saving message so the status bar does not appear to blink on
and off.

The NuGet addin still saves the project more often than it needs to.
The original implementation was based on how Visual Studio works, which
saves the project on adding a reference, adding an import or removing
an import. However saving the project once at the end may be a
performance improvement that can be looked at in the future.