[Ide] Prevent language keyword being used as new project name.
Fixed bug #15478 - MonoDevelop shouldn't accept language keywords as
Project Name
https://bugzilla.xamarin.com/show_bug.cgi?id=15478

If a new C# project was created with the name 'try' then the project
would not compile since the generated files would have an invalid
namespace of 'try'. To prevent this the new project dialog now checks
the CodeDomProvider for the template's language to see if the project
name is a valid identifier. If this is not the case an error message
is displayed and the project cannot be created until its name is
changed.

Another approach would be to escape the namespace in the
SingleFileDescriptionTemplate using the CodeDomProvider's
CreateEscapedIdentifier. Whilst that works for C# projects it
does not work for F# projects. If you try to create an F# project
named 'try' it does not escape the namespace or use the namespace
created by the IDE.