[NuGet] Stop using DispatcherService.
The NuGet addin no longer uses the DispatcherService to run package
actions in the background. It now uses its own background thread to
queue and run these actions.

Due to being unable to use the original DispatcherService's
BackgroundDispatchAndWait without locking up the UI the
PackageManagementProjectOperations.InstallPackages has been changed
to return when the package install has started. Previously it would
wait until the packages were finished installing.

Also the PackageManagementProjectOperations.GetIstalledPackages now
returns installed packages as well as those being installed. This
prevents a Xamarin Component that installs a NuGet package from
showing an error when it is installed. Unfortunately one side effect
of this change is that if the NuGet package fails to install the
Component does not show an error icon when previously it did since
originally the packages were installed and then the list of packages
was checked. Now this happens whilst the packages are being installed.