[Ide] Fix memory leak in New Project dialog.
Depending on which page was shown on closing the dialog either the
UI control for the project preview page or the templates page was
not destroyed on closing the dialog. The TreeSelection.SelectFunction
methods would then keep a reference to the dialog since the function
wrappers were still referenced on the heap.