[Ide] Improved handling of files outside of projects.
It doesn't make sense to show errors related to resolving or to
provide code issues for them. (Due to resolve errors many code issues
won't work)
So resolve errors & code issues are turned off and the rest is
working. Which means that basic semantic highlighting, code completion
and other parse errors are unaffected - it's just easier to work with
complex stand alone files this way.