[Ide] Handle directory separator chars when renaming files and folders
Fixed bug #37309 - Creating project new folder with '\' at the end
of name causes endless loop recursion
https://bugzilla.xamarin.com/show_bug.cgi?id=37309

If a folder is renamed and a '\' char is used at the end then on
Windows the folder would be renamed and the Solutions window would
show a folder item that could be expanded repeatedly, with the same
folder inside it. Closing and re-opening the solution fixes the
problem. If a directory separator character is used in the middle of
the folder name then an error occurs when an attempt is made to create
the directory.

Now when renaming a file or a folder if a directory separator char
is used then a message will be displayed indicating that the name is
invalid. This is already handled with projects, solutions and
workspaces in the Solution window. Now it is also handled with files
and folders.